Fixed bugs with SQL data being read on a different type

+++ Server/Product.cs
@@ -9,9 +9,9 @@ namespace Server
         public string id { get; set; }
         public string itemname { get; set; }
         public string description { get; set; }
-        public float itemprice { get; set; }
+        public decimal itemprice { get; set; }
         public string[] pictureids { get; set; }
-        public float discountpercent { get; set; }
+        public decimal discountpercent { get; set; }
         public string[] categoriesid { get; set; }
         public string shopid { get; set; }
